set Box as deprecated
