+++ src/components/box/_box.scss
@@ -5,7 +5,7 @@ $boxBorderSize: 1px;
 $includeHtml: false !default;
 
 @if ($includeHtml) {
-  .sg-box {
+  .sg-box-deprecated {
     @include component();
     @include hole();
     display: inline-flex;
@@ -26,7 +26,7 @@ $includeHtml: false !default;
     &--full {
       width: 100%;
 
-      .sg-box__hole {
+      .sg-box-deprecated__hole {
         width: 100%;
         max-width: 100%;
       }
