+++ src/docs/page-components.jsx
@@ -9,6 +9,7 @@ import FlashMessages from 'flash-messages/pages/flash-messages-interactive';
 import SubjectIcons from 'subject-icons/pages/subject-icons-interactive';
 import Subjects from 'subject-icons/pages/subjects-interactive';
 import IconsAsButtons from 'icon-as-button/pages/icon-as-button-interactive';
+import BoxesDeprecated from 'box/pages/box-deprecated-interactive';
 import Boxes from 'box/pages/box-interactive';
 import Flexboxes from 'flex/pages/flex-interactive';
 import Cards from 'card/pages/card-interactive';
@@ -54,6 +55,7 @@ const demos = {
   'Subject icon box': <Subjects />,
   'Flash messages': <FlashMessages />,
   Boxes: <Boxes />,
+  BoxesDeprecated: <BoxesDeprecated />,
   Flex: <Flexboxes />,
   Cards: <Cards />,
   'Action List': <ActionLists />,
